converted float to double prec. Should prevent long term precision errors.

+++ Env.h
@@ -11,15 +11,15 @@ private:
 // arrays for ATOMS
 std::vector<std::string> Aatom_name;
 std::vector<std::string> Aresidue;
-//std::vector<float> Acoords;
+//std::vector<double> Acoords;
 
 // arrays for HETATMS
 
 std::vector<std::string> Hatom_name;
 std::vector<std::string> Hresidue;
-std::vector<float> Hcoords;
+std::vector<double> Hcoords;
 public:
-std::vector<float> Acoords;
+std::vector<double> Acoords;
 Environment(std::string pdb);
 void append_name(std::string name);
 void append_Hname(std::string name);
