+++ Vector_Math.cpp
@@ -2,7 +2,7 @@
 #include "Env.h"
 #include <math.h>
 
-void vect_add(std::vector<float>& v1, std::vector<float>& v2, std::vector<float>& product){
+void vect_add(std::vector<double>& v1, std::vector<double>& v2, std::vector<double>& product){
 
     for(int i = 0; i < v1.size(); i++){
         product[i] = v1[i] + v2[i]; 
@@ -10,14 +10,14 @@ void vect_add(std::vector<float>& v1, std::vector<float>& v2, std::vector<float>
 
 }
 
-void resize(std::vector<float>& vect, float scale){
+void resize(std::vector<double>& vect, double scale){
     for(int i = 0; i < vect.size(); i++){
         vect[i] *= scale;
     }
 }
 
-void magnitude(std::vector<float>& object, float& mag){
-    float temp = 0;
+void magnitude(std::vector<double>& object, double& mag){
+    double temp = 0;
     for(int i = 0; i < object.size(); i++){
         temp+= object[i] *object[i];
     }
@@ -25,26 +25,26 @@ void magnitude(std::vector<float>& object, float& mag){
     mag = temp;
 }
 
-void displacement_vect(std::vector<float>& d, std::vector<float>& Acoords , int atom1, int atom2){
+void displacement_vect(std::vector<double>& d, std::vector<double>& Acoords , int atom1, int atom2){
     for(int i = 0; i < 3; i++){
         d.push_back(Acoords[atom1 * 3 + i] - Acoords[atom2 * 3 + i]);
     }
 }
 
-void unit_vector(float& mag, std::vector<float> d, std::vector<float>& unitv){
+void unit_vector(double& mag, std::vector<double> d, std::vector<double>& unitv){
     
     for(int i = 0; i < d.size(); i++){
         unitv.push_back(d[i]/mag);
     }
 }
 
-void dot(std::vector<float>& disp1, std::vector<float>& disp2, float& val){
+void dot(std::vector<double>& disp1, std::vector<double>& disp2, double& val){
     for(int i = 0; i < 3; i++){
         val += disp1[i] * disp2[i];
     }
 }
 
-void cross(std::vector<float>& vect1, std::vector<float>& vect2, std::vector<float>& cprod){
+void cross(std::vector<double>& vect1, std::vector<double>& vect2, std::vector<double>& cprod){
 
     cprod.push_back(vect1[1] * vect2[2] - vect1[2] * vect2[1]);
     cprod.push_back(-(vect1[0] * vect2[2] - vect1[2] * vect2[0]));
